% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mutation_rates.R
\name{adjust_indel_rates}
\alias{adjust_indel_rates}
\title{adapt indel rates for lower rate estimate from validated de novos}
\usage{
adjust_indel_rates(rates)
}
\arguments{
\item{rates}{data frame of mutation rates.}
}
\value{
the rates data frame, with adjusted indel rates.
}
\description{
The indel mutation rates from Samocha et al., Nature Genetics 46:944-950
assume that the overall indel mutation rate is 1.25-fold greater than the
overall nonsense mutation rate, ie there are 1.25 times as many frameshifts
as nonsense mutations. We have our own estimates for the ratio, derived from
our de novo validation efforts, which we shall apply in place of the Samocha
et al ratios.
}
\examples{
rates = read.table(header=TRUE, text="
    hgnc    chrom  snv.missense.rate  snv.lof.rate  indel.missense.rate  indel.lof.rate
    ARID1B  6      0.001              0.001         0.001                0.001
    KMT2A   11     0.001              0.001         0.001                0.001
    AMELX   X      0.001              0.001         0.001                0.001")
adjust_indel_rates(rates)
}

