% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mupit_core.R
\name{analyse_gene_enrichment}
\alias{analyse_gene_enrichment}
\title{analyse whether de novo mutations are enriched in genes}
\usage{
analyse_gene_enrichment(de_novos, trios, plot_path = NA, all_genes = FALSE,
  rates = NULL)
}
\arguments{
\item{de_novos}{data frame containing all the observed de novos for all the
genes}

\item{trios}{list of male and female proband counts in the population}

\item{plot_path}{path to save enrichment plots to}

\item{all_genes}{whether to test all genes in the genome (most will test
the probability of observing 0 de novos).}

\item{rates}{gene-based mutation rates data frame, or NA}
}
\value{
data frame containing results from testiong for enrichment of de
    in each gene with de novos in it.
}
\description{
analyse whether de novo mutations are enriched in genes
}
\examples{
trios = list(male=1000, female=1000)
vars = read.table(header=TRUE, text="
     person_id hgnc chrom start_pos consequence type
     person_1 ARID1B 6 157150547 inframe_deletion indel
     person_2 ARID1B 6 157431695 missense_variant snv
     person_3 ARID1B 6 157454186 frameshift_variant indel
     person_4 ARID1B 6 157502190 stop_gained snv
     person_4 KMT2A 11 118367083 missense_variant snv")
rates = read.table(header=TRUE, text="
     hgnc    chrom  syn  mis  non   splice_site  frameshift
     ARID1B  6      -5   -6   -6.5  -5           -7
     KMT2A   11     -5   -6   -6.5  -5           -7")
analyse_gene_enrichment(vars, trios, rates=rates)
}

