% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cohorts}
\alias{cohorts}
\title{Cohort information.}
\format{A data frame with eight variables: \code{study_code}, \code{year},
    \code{study_phenotype}, \code{publication_doi}, \code{unique_male},
    \code{unique_female}, \code{subsets}, and \code{comment}.}
\usage{
cohorts
}
\description{
Table of sex-specific numbers of affected probands reported on in published
exome sequencing of children with developmental disorders (autism, congenital
heart disease, intellectual disability, epilepsy, and schizophrenia).
}
\details{
This provides the total population of trio-based probands investigated within
these studies, that is, the population for whom a de novo mutation might have
been identified. This determines the number of mutations we expect to find
within the total population.

The table of sex-counts has been carefully extracted from the publications,
accounting for overlap in samples between different publications.
}
\keyword{datasets}

