% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combine_analyses.R
\name{combine_enrichment_and_clustering}
\alias{combine_enrichment_and_clustering}
\title{combine P values from enrichment and clustering tests into a single P value}
\usage{
combine_enrichment_and_clustering(enriched, clust, num_tests = 18500)
}
\arguments{
\item{enriched}{dataframe of de novo enrichment results}

\item{clust}{dataframe of de novo clustering results}

\item{num_tests}{number of tests to correct for multiple testing}
}
\value{
a merged dataset where the P values have been combined
}
\description{
combine P values from enrichment and clustering tests into a single P value
}

