% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combine_analyses.R
\name{combine_tests}
\alias{combine_tests}
\title{find the most significant P value for each gene from the P values from
different subsets and different tests}
\usage{
combine_tests(meta_clust, meta_enrich, clust, enrich, pheno_path = NULL)
}
\arguments{
\item{meta_clust}{path to clustering results for the meta-analysis subset}

\item{meta_enrich}{path to enrichment results for the meta-analysis subset}

\item{clust}{path to clustering results for the ddd only subset}

\item{enrich}{path to enrichment results for the ddd only subset}

\item{pheno_path}{path to phenotype similarity testing results}
}
\value{
data frame with the columns from all the datasets, as well as minimum
    P values from each subset for each gene, and overall minimum P values for
    each gene.
}
\description{
find the most significant P value for each gene from the P values from
different subsets and different tests
}

