% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutation_rates.R
\name{correct_for_x_chrom}
\alias{correct_for_x_chrom}
\title{correct mutations rates for sex-chromosome transmission rates}
\usage{
correct_for_x_chrom(rates, male_n, female_n)
}
\arguments{
\item{rates}{gene-based data frame, containing rates for different mutation
classes.}

\item{male_n}{number of trios with male offspring}

\item{female_n}{number of trios with female offspring}
}
\value{
a dataframe of mutation rates for genes under different mutation
    classes.
}
\description{
correct mutations rates for sex-chromosome transmission rates
}
\examples{
rates = read.table(header=TRUE, text="
     hgnc    chrom  snv.missense.rate  snv.lof.rate  indel.missense.rate  indel.lof.rate
     ARID1B  6      0.001              0.001         0.001                0.001
     KMT2A   11     0.001              0.001         0.001                0.001
     AMELX   X      0.001              0.001         0.001                0.001")
correct_for_x_chrom(rates, 1000, 1000)
}

