% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ddd_de_novos}
\alias{ddd_de_novos}
\title{De novos found in the DDD study.}
\format{A data frame with twelve variables: \code{person_id}, \code{chrom},
    \code{start_pos}, \code{end_pos}, \code{ref_allele}, \code{alt_allele},
    \code{hgnc}, \code{consequence}, \code{study_code},
    \code{publication_doi}, \code{study_phenotype} and \code{type}}
\usage{
ddd_de_novos
}
\description{
De novo variants found within participants in the DDD study. As this is an
unpblished dataset, the dataset in the package will be a blank dataframe,
and only be complete for DDD researchers.
}
\keyword{datasets}

