% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ensembl_rest.R
\name{find_most_severe_transcript}
\alias{find_most_severe_transcript}
\title{find the most severe transcript from Ensembl data}
\usage{
find_most_severe_transcript(ensembl_json, include_hgvs = FALSE,
  exclude_bad = TRUE)
}
\arguments{
\item{ensembl_json}{json data for variant from Ensembl}

\item{include_hgvs}{boolean showing if we will be using HGVS predictions from
the VEP consequence predictions. If true, we sort the trasncripts by their
length, so that we consistently return the same transcript between
different variants from the same gene. Otherwise we just return the first
most severe transcript.}

\item{exclude_bad}{boolean to show whether we want to exclude nonsense
    mediated decay transcripts and so forth. We only supply a false value for
    recursive calling, when we have failed to find any transcript during
    normal usage.}
}
\value{
a character string containing the most severe consequence, as per VEP
    annotation formats.
}
\description{
find the most severe transcript from Ensembl data
}

