% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/variant_cleaning.R
\name{fix_coordinates_with_allele}
\alias{fix_coordinates_with_allele}
\title{extract genomic coordinates for variants from a dataframe}
\usage{
fix_coordinates_with_allele(variants, coordinate_column, allele_column)
}
\arguments{
\item{variants}{data frame of variants}

\item{coordinate_column}{name of column containing chromosome information}

\item{allele_column}{name of column containing allele information}
}
\value{
a data frame with chrom, start_pos, end_pos and allele columns.
}
\description{
Genomic coordinates can be encoded in a variety of ways, this function
cleans up coordinates where one column
}
\examples{
fix_coordinates_with_allele(data.frame(coord=c("chr1:10000"),
    allele=c("A/G")), "coord", "allele")
fix_coordinates_with_allele(data.frame(coord=c("chr1:10000"),
    allele=c("ATGC/G")), "coord", "allele")
fix_coordinates_with_allele(data.frame(coord=c("chr1:10000"),
    allele=c("sub(G-&gt;T)")), "coord", "allele")
fix_coordinates_with_allele(data.frame(coord=c("chr1:10000"),
    allele=c("del(1)")), "coord", "allele")
fix_coordinates_with_allele(data.frame(coord=c("chr1:10000"),
    allele=c("ins(ATG)")), "coord", "allele")
}

