% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/variant_cleaning.R
\name{fix_coordinates_with_hgvs_genomic}
\alias{fix_coordinates_with_hgvs_genomic}
\title{extract genomic coordinates for variants encoded as HGVS genomic}
\usage{
fix_coordinates_with_hgvs_genomic(variants, hgvs_column)
}
\arguments{
\item{variants}{data frame of variants}

\item{hgvs_column}{name of column containing HGVS coordinate}
}
\value{
a data frame with chrom, start_pos, end_pos and allele columns.
}
\description{
extract genomic coordinates for variants encoded as HGVS genomic
}
\examples{
fix_coordinates_with_hgvs_genomic(
   data.frame(hgvs=c("chr7:g.155556643G>A")), "hgvs")
fix_coordinates_with_hgvs_genomic(data.frame(
   hgvs=c("chr3:g.11060365_11060365del")), "hgvs")
fix_coordinates_with_hgvs_genomic(data.frame(
   hgvs=c("chr13:g.50057690_50057691insA")), "hgvs")
}

