% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fix_alleles.R
\name{fix_del_alleles}
\alias{fix_del_alleles}
\title{fix deletion allele codes}
\usage{
fix_del_alleles(variants, allele_column)
}
\arguments{
\item{variants}{data frame of variants}

\item{allele_column}{name of column containing allele information}
}
\value{
a data frame with chrom, start_pos, end_pos and allele columns.
}
\description{
fix the allele column for variants with allele columns that are structured
like "del(1)"
}

