% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fix_alleles.R
\name{fix_het_alleles}
\alias{fix_het_alleles}
\title{correct alt alleles which have been encoded as an IUPAC ambiguous base}
\usage{
fix_het_alleles(variants)
}
\arguments{
\item{variants}{data frame of variants}
}
\value{
a data frame with chrom, start_pos, end_pos and allele columns.
}
\description{
Sometimes the de novo variants from studies provide alt alleles as "R", or
"Y", which indicate ambigous bases. We can identify the correct alt base by
comparison with the reference allele.
}

