% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fix_alleles.R
\name{fix_sub_alleles}
\alias{fix_sub_alleles}
\title{fix substitution allele codes}
\usage{
fix_sub_alleles(variants, allele_column)
}
\arguments{
\item{variants}{data frame of variants}

\item{allele_column}{name of column containing allele information}
}
\value{
a data frame with chrom, start_pos, end_pos and allele columns.
}
\description{
fix the allele column for variants with allele columns that are structured
like "sub(G-&gt;T)" (the "-&gt;" translates to ">", and sub(G->T), this is
an excel to R conversion issue. Perhaps it is unicode from excel?
}

