% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/variant_cleaning.R
\name{fix_zaiidi_coordinates}
\alias{fix_zaiidi_coordinates}
\title{extract genomic coordinates for variants from a dataframe}
\usage{
fix_zaiidi_coordinates(variants, allele_column)
}
\arguments{
\item{variants}{data frame of variants}

\item{allele_column}{name of column containing allele information}
}
\value{
a data frame with chrom, start_pos, end_pos and allele columns.
}
\description{
Genomic coordinates can be encoded in a variety of ways, this function
cleans up coordinates where one column
}
\examples{
fix_zaiidi_coordinates(data.frame(chrom=c("1"), start_pos=c("100000"),
    alleles=c("A/G")), "alleles")
fix_zaiidi_coordinates(data.frame(chrom=c("1"), start_pos=c("100000"),
    alleles=c("-AAAA")), "alleles")
fix_zaiidi_coordinates(data.frame(chrom=c("1"), start_pos=c("100000"),
    alleles=c("+AAAA")), "alleles")
}

