% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mupit_core.R
\name{gene_enrichment}
\alias{gene_enrichment}
\title{tests whether genes are enriched with de novo mutations}
\usage{
gene_enrichment(rates, counts, all_genes = FALSE)
}
\arguments{
\item{rates}{gene mutation rates per consequence type}

\item{counts}{data frame with tally of de novo mutations per gene for each of
the mutation types: lof_snv, lof_indel, missense_snv, missense_indel.}

\item{all_genes}{whether to test all genes in the genome (most will test
    the probability of observing 0 de novos).}
}
\value{
data frame with gene info, mutation rates and P values from testing
    for enrichment.
}
\description{
tests whether genes are enriched with de novo mutations
}
\examples{
counts = read.table(header=TRUE, text="
    hgnc   lof_indel lof_snv missense_indel missense_snv chrom min_pos
    ARID1B 1       1            1         1              6     157150547
    KMT2A  0       0            0         1              11    118367083")

rates = read.table(header=TRUE, text="
    hgnc   chrom snv.lof.rate indel.lof.rate snv.missense.rate indel.missense.rate
    ARID1B 6     0.01        0.005          0.05             0.005
    KMT2A  11    0.01        0.005          0.05             0.005")

gene_enrichment(rates, counts)
}

