% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_info}
\alias{gene_info}
\title{Gene information.}
\format{A data frame with three variables: \code{hgnc}, \code{chrom},
    and \code{cds_length}}
\source{
url(http://www.ensembl.org/biomart)
}
\usage{
gene_info
}
\description{
Table of all genes in the genome with HGNC symbol, chromosome and coding
sequence (CDS) length in base-pairs.
}
\keyword{datasets}

