% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/open_ddd_data.R
\name{get_clinical_filtering_vcf_paths}
\alias{get_clinical_filtering_vcf_paths}
\title{find paths to clinical filtering output for DDD samples}
\usage{
get_clinical_filtering_vcf_paths(proband_ids)
}
\arguments{
\item{proband_ids}{vector of proband IDs}
}
\value{
vector of VCF paths
}
\description{
find paths to clinical filtering output for DDD samples
}

