% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/open_ddd_data.R
\name{get_ddd_de_novos}
\alias{get_ddd_de_novos}
\title{get de novo data for DDD study.}
\usage{
get_ddd_de_novos(diagnosed = NULL, subset = NULL)
}
\arguments{
\item{diagnosed}{dataframe of samples in DDD with a diagnosis, who should be
excluded from analyses}

\item{subset}{vector of sample IDs, if we want to restrict to a specific
        set of probands, such as when we are analysing a phenotype-specific
        subset of the DDD.}
}
\value{
data frame of de novos, including gene symbol, functional consequence
    (VEP format), chromosome, nucleotide position and SNV or INDEL type
}
\description{
get de novo data for DDD study.
}

