% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/open_ddd_data.R
\name{get_ddd_diagnosed}
\alias{get_ddd_diagnosed}
\title{find diagnosed probands in the DDD study, to exclude them from our data}
\usage{
get_ddd_diagnosed(path)
}
\arguments{
\item{path}{path to file defining diagnosed probands}
}
\value{
A list containing vectors with DDD IDs, and sex of the diagnosed
    probands
}
\description{
find diagnosed probands in the DDD study, to exclude them from our data
}

