% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mupit_core.R
\name{get_de_novo_counts}
\alias{get_de_novo_counts}
\title{tallies the mutation types observed for each gene}
\usage{
get_de_novo_counts(de_novos)
}
\arguments{
\item{de_novos}{data frame listing all the de novo mutations, with columns
for HGNC symbol, consequence type (VEP style predictions), and a
column indicating SNV, or indel.}
}
\value{
data frame with tally of de novo mutations for each of the mutation
        types
}
\description{
tallies the mutation types observed for each gene
}
\examples{
vars = read.table(header=TRUE, text="
   person_id   hgnc     chrom   start_pos   consequence        type
   person_1    ARID1B   6       157431695   missense_variant   snv
   person_2    ARID1B   6       157502190   stop_gained        snv")
get_de_novo_counts(vars)
}

