% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutation_rates.R
\name{get_gene_based_mutation_rates}
\alias{get_gene_based_mutation_rates}
\title{gets mutation rates from the gene-based datasets}
\usage{
get_gene_based_mutation_rates(trios, rates = NULL)
}
\arguments{
\item{trios}{list of male and female proband counts in the dataset}

\item{rates}{optionally specify a dataframe containing per-gene mutation
rates, or NULL}
}
\value{
a dataframe of mutation rates for genes under different mutation
    classes.
}
\description{
This defaults to the gene-based mutation rates from Nature Genetics
46:944-950 (2014) doi:10.1038/ng.3050, but we can pass in other gene-based
mutation rate datasets.
}
\examples{
trios = list(male=1000, female=1000)
rates = read.table(header=TRUE, text="
    hgnc    chrom  syn  mis  non   splice_site  frameshift
    ARID1B  6      -5   -6   -6.5  -5           -7
    KMT2A   11     -5   -6   -6.5  -5           -7
    AMELX   X      -5   -6   -6.5  -5           -7")
get_gene_based_mutation_rates(trios)
get_gene_based_mutation_rates(trios, rates)
}

