% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_vcf.R
\name{get_hgnc_from_vcf_info}
\alias{get_hgnc_from_vcf_info}
\title{get HGNC symbols from the INFO field of VCF variants}
\usage{
get_hgnc_from_vcf_info(info)
}
\arguments{
\item{info}{vector of VCF INFO entries for different variants}
}
\value{
vector of HGNC symbols, in same order as the INFO vector
}
\description{
get HGNC symbols from the INFO field of VCF variants
}

