% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mutation_rates.R
\name{get_length_based_rates}
\alias{get_length_based_rates}
\title{calculate mutation rates based on the gene length}
\usage{
get_length_based_rates(trios)
}
\arguments{
\item{trios}{list of male and female proband counts in the dataset}
}
\value{
a dataframe of mutation rates for genes under different mutation
    classes.
}
\description{
This has been discontinued in favor of using rates derived from de novo
mutation rates, from Nature Genetics 46:944-950.
}

