% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/open_ddd_data.R
\name{get_likely_diagnosed}
\alias{get_likely_diagnosed}
\title{find probands likely to have diagnoses, to exclude them from our data}
\usage{
get_likely_diagnosed(path)
}
\arguments{
\item{path}{path to file defining diagnosed probands}
}
\value{
A list containing vectors with DDD IDs, and sex of the diagnosed
    probands
}
\description{
We assume DDD samples that have
}

