% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ensembl_rest.R
\name{get_protein_length}
\alias{get_protein_length}
\title{find the coding length for a ensembl transcript}
\usage{
get_protein_length(transcript_id, build = "grch37", verbose = FALSE)
}
\arguments{
\item{transcript_id}{json data for variant from Ensembl}

\item{build}{genome build to find consequences on}

\item{verbose}{flag indicating whether to print variants as they are checked}
}
\value{
integer for CDS length.
}
\description{
find the coding length for a ensembl transcript
}

