% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_vcf.R
\name{get_variants}
\alias{get_variants}
\title{open all the variants found within a set of single-sample VCFs}
\usage{
get_variants(vcf_paths)
}
\arguments{
\item{vcf_paths}{vector of paths to VCFs}
}
\value{
dataframe of all variants in the VCFs
}
\description{
open all the variants found within a set of single-sample VCFs
}

