% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ensembl_rest.R
\name{get_vep_consequence}
\alias{get_vep_consequence}
\title{find the VEP consequence for a variant}
\usage{
get_vep_consequence(variant, include_hgvs = FALSE,
  include_deleterious_score = FALSE, build = "grch37", verbose = FALSE)
}
\arguments{
\item{variant}{data frame or list for a variant, containing columns named
"chrom", "start_pos", "end_pos", and "alt_allele" code for a single variant}

\item{include_hgvs}{whether to also include the HGVS consequence strings for
cDNA and protein.}

\item{include_deleterious_score}{whether to also include SIFT and polyphen predictions}

\item{build}{genome build to find consequences on}

\item{verbose}{flag indicating whether to print variants as they are checked}
}
\value{
a character string containing the most severe consequence, as per VEP
    annotation formats.
}
\description{
find the VEP consequence for a variant
}
\examples{
get_vep_consequence(data.frame(chrom=c("1"),
    start_pos=c("1000000"), end_pos=c("1000000"), alt_allele=c("A"),
    ref_allele=c("G")))
get_vep_consequence(list(chrom="1", start_pos="1000000",
    end_pos="1000000", alt_allele="A", ref_allele="G"))
}

