% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enrichment_plot_functions.R
\name{label_genes}
\alias{label_genes}
\title{make plot labels for genes with fdr > threshold}
\usage{
label_genes(enriched, p_values, num_tests)
}
\arguments{
\item{enriched}{data frame containing HGNC symbols}

\item{p_values}{vector of p-values, sorted as per enriched data frame}

\item{num_tests}{number of tests performed (used for multiple correction).}
}
\description{
make plot labels for genes with fdr > threshold
}

