% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/open_ddd_data.R
\name{load_ddg2p}
\alias{load_ddg2p}
\title{load the current DDG2P dataset.}
\usage{
load_ddg2p(path)
}
\arguments{
\item{path}{path to DDD de novo dataset}
}
\value{
data frame of DDG2P genes
}
\description{
load the current DDG2P dataset.
}

