% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_vcf.R
\name{load_vcf}
\alias{load_vcf}
\title{open all the variants found within a single-sample VCF}
\usage{
load_vcf(path)
}
\arguments{
\item{path}{path to VCF}
}
\value{
dataframe of all variants in the VCF, modified so that the sample ID
    is an additional column
}
\description{
open all the variants found within a single-sample VCF
}

