% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mupit-package.R
\docType{package}
\name{mupit}
\alias{mupit}
\alias{mupit-package}
\title{mupit: examines enrichment of de novo mutations within genes.}
\description{
mupit: examines enrichment of de novo mutations within genes.
}

