% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enrichment_plot_functions.R
\name{plot_enrichment_graphs}
\alias{plot_enrichment_graphs}
\title{make Manhattan plots for LOF and Func variants separately}
\usage{
plot_enrichment_graphs(enriched, num_tests, path)
}
\arguments{
\item{enriched}{data frame containing columns for chr, coord (position), and
p values from testing for enrichment with loss-of-function and
functional consequence variants.}

\item{num_tests}{number of tests performed (used by Bonferroni and FDR
correction).}

\item{path}{path to save the plot to}
}
\description{
make Manhattan plots for LOF and Func variants separately
}

