% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enrichment_plot_functions.R
\name{plot_values}
\alias{plot_values}
\title{make Manhattan plots for LoF, func and silent variants separately}
\usage{
plot_values(num_tests, enriched, p_values, title)
}
\arguments{
\item{num_tests}{number of tests performed (used by Bonferroni and FDR
correction).}

\item{enriched}{data frame containing HGNC symbols}

\item{p_values}{vector of P values from enrichment testing}

\item{title}{character string for the title of the plot}
}
\description{
make Manhattan plots for LoF, func and silent variants separately
}

