% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ensembl_rest.R
\name{request_from_ensembl}
\alias{request_from_ensembl}
\title{make a URL request to the Ensembl service}
\usage{
request_from_ensembl(ext, headers = "", build = "grch37", tries = 0)
}
\arguments{
\item{ext}{extension for URL}

\item{headers}{headers for the url}

\item{build}{genome build for REST API}

\item{tries}{number of attempts that have been mader to access the URL}
}
\value{
a character string, typically json encoded
}
\description{
make a URL request to the Ensembl service
}
\examples{
request_from_ensembl("/vep/human/region/1:205901016:205901016/A")
}

