% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/open_ddd_data.R
\name{standardise_ddd_de_novos}
\alias{standardise_ddd_de_novos}
\title{get standardised de novo data for DDD study.}
\usage{
standardise_ddd_de_novos()
}
\value{
data frame of de novos, including gene symbol, functional consequence
    (VEP format), chromosome, nucleotide position and SNV or INDEL type
}
\description{
get standardised de novo data for DDD study.
}

