% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/write_json_probands.R
\name{write_probands_by_gene}
\alias{write_probands_by_gene}
\title{Write a json-formatted list of probands per gene}
\usage{
write_probands_by_gene(de_novos, path)
}
\arguments{
\item{de_novos}{dataframe of de-novo mutations per proband}

\item{path}{path to write the json data to}
}
\description{
We want a correctly-formatted JSON list of probands per gene, for analysis of
phenotype similarity between probands, see
\url{https://github.com/jeremymcrae/hpo_similarity}.
}

