% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/combine_analyses.R
\name{combine_enrichment_and_clustering}
\alias{combine_enrichment_and_clustering}
\title{combine P values from enrichment and clustering tests into a single P value}
\usage{
combine_enrichment_and_clustering(enrichment_path, clustering_path)
}
\arguments{
\item{enrichment_path}{path to enrichment testing results}

\item{clustering_path}{path to clustering testing results}
}
\value{
a merged dataset where the P values have been combined
}
\description{
combine P values from enrichment and clustering tests into a single P value
}

