% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/combine_analyses.R
\name{fishersMethod}
\alias{fishersMethod}
\title{function to combine p values, using Fisher's method}
\usage{
fishersMethod(x)
}
\arguments{
\item{x}{vector of P values for a gene}
}
\value{
single P value for gene
}
\description{
function to combine p values, using Fisher's method
}

